+++ loafer/exceptions.py
@@ -3,6 +3,10 @@ class ProviderError(Exception):
     pass
 
 
+class ProviderRuntimeError(ProviderError):
+    pass
+
+
 class ConfigurationError(Exception):
     pass
 
@@ -12,8 +16,4 @@ class LoaferException(Exception):
 
 
 class DeleteMessage(LoaferException):
-    pass
-
-
-class SQSProviderRuntimeError(RuntimeError):
     pass